- Korasi's Sword added - Xamphur drop table added - Sigil of deft strike accuracy adjustment fixed - fixed drop rates with pvp equipment from slayer

+++ src/main/java/com/cryptic/model/entity/combat/damagehandler/impl/armor/DragonHunterLance.java
@@ -44,7 +44,7 @@ public class DragonHunterLance implements DamageModifyingListener {
                         if (FormulaUtils.isDragon(npc)) {
                             var modifier = accuracy.modifier();
                             modifier += 1.20;
-                            return modifier;
+                            return (int) modifier;
                         }
                     }
                 }
