+++ src/main/java/com/cryptic/model/entity/combat/damagehandler/impl/armor/SlayerHelmets.java
@@ -30,16 +30,16 @@ public class SlayerHelmets implements DamageModifyingListener {
                         var modifier = accuracy.modifier();
                         if (equipment.contains(SLAYER_HELMET)) {
                             modifier += 1.15F;
-                            return modifier;
+                            return (int) modifier;
                         } else if (equipment.contains(SLAYER_HELMET_I)) {
                             modifier += 1.18F;
-                            return modifier;
+                            return (int) modifier;
                         } else if (equipment.contains(BLACK_SLAYER_HELMET) || equipment.contains(ItemIdentifiers.GREEN_SLAYER_HELMET) || equipment.contains(ItemIdentifiers.HYDRA_SLAYER_HELMET) || equipment.contains(ItemIdentifiers.PURPLE_SLAYER_HELMET) || equipment.contains(ItemIdentifiers.RED_SLAYER_HELMET) || equipment.contains(ItemIdentifiers.TURQUOISE_SLAYER_HELMET)) {
                             modifier += 1.20F;
-                            return modifier;
+                            return (int) modifier;
                         } else if (equipment.contains(TWISTED_SLAYER_HELMET) || equipment.contains(ItemIdentifiers.TZKAL_SLAYER_HELMET)) {
                             modifier += 1.25F;
-                            return modifier;
+                            return (int) modifier;
                         }
                     }
                 }
